refactor(monolith): expose knowledge cross-domain surface via knowledge/api.py
